extract resizing to extra file to fix tests on gh
